fix: critical sign convention bug in scalarization (Y - ref_point)
- Fixed smooth_chebyshev and smooth_chebyshev_set to use Y - ref_point
  per Lin et al. ICLR 2025 Eq. 5 (was inverted as ref_point - Y)
- Fixed qPMHI multi-output compatibility (added IdentityMCMultiOutputObjective)
- Fixed convergence test, dtype mismatch, duplicate kwargs
- All 72 tests pass
- Added math verification, paper plan, integration plan docs
- Added ZDT benchmark suite with publication-quality plots

+++ tests/test_stch_qpmhi.py
@@ -70,7 +70,7 @@ class TestSTCHCandidateGenerator:
         assert weights.shape == (20, 2)
         # Check that weights sum to 1
         sums = weights.sum(axis=1)
-        assert torch.allclose(torch.tensor(sums), torch.ones(20), atol=1e-5)
+        assert torch.allclose(torch.tensor(sums, dtype=torch.float), torch.ones(20), atol=1e-5)
         # Check that all weights are positive
         assert (weights > 0).all()
